--- app/src/main/java/com/example/futureinsight/MainActivity.java
@@ -1,33 +1,0 @@
-package com.example.futureinsight;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.os.Handler;
-import android.view.WindowManager;
-
-public class MainActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        try {
-            super.onCreate(savedInstanceState);
-            getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN,WindowManager.LayoutParams.FLAG_FULLSCREEN);
-            setContentView(R.layout.activity_main);
-            new Handler().postDelayed(new Runnable() {
-                @Override
-                public void run() {
-                    Intent i = new Intent(MainActivity.this, HomeActivity.class);
-                    startActivity(i);
-                    finish();
-                }
-            },2000);
-
-        }catch (Exception e) {
-            System.out.println("Exception: " + e);
-            System.out.println("Exception: " + e.getMessage());
-            System.out.println("Application is Down");
-        }
-    }
-}
