+++ app/src/main/res/layout/fragment_gallery.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.gallery.GalleryFragment"
+    tools:context="com.futureinsight.futureinsight.ui.gallery.GalleryFragment"
     android:background="@color/black">
 
     <WebView
