+++ app/src/main/res/layout/fragment_slideshow.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.slideshow.SlideshowFragment"
+    tools:context="com.futureinsight.futureinsight.ui.slideshow.SlideshowFragment"
     android:background="@color/black">
 
     <WebView
