+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,20 +7,20 @@
 
     <fragment
         android:id="@+id/nav_home"
-        android:name="com.example.futureinsight.ui.home.HomeFragment"
+        android:name="com.futureinsight.futureinsight.ui.home.HomeFragment"
         android:label="@string/menu_home"
         tools:layout="@layout/fragment_home"
         />
 
     <fragment
         android:id="@+id/nav_gallery"
-        android:name="com.example.futureinsight.ui.gallery.GalleryFragment"
+        android:name="com.futureinsight.futureinsight.ui.gallery.GalleryFragment"
         android:label="@string/blogs"
         tools:layout="@layout/fragment_gallery"
         />
     <fragment
         android:id="@+id/nav_slideshow"
-        android:name="com.example.futureinsight.ui.slideshow.SlideshowFragment"
+        android:name="com.futureinsight.futureinsight.ui.slideshow.SlideshowFragment"
         android:label="@string/services"
         tools:layout="@layout/fragment_slideshow"
         />
